Getting the "spaceball" demo to build properly in Windows
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@844 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/progs/demos/spaceball/vmath.c
@@ -4,10 +4,10 @@
 quat_t quat_rotate(quat_t q, float angle, float x, float y, float z)
 {
   quat_t rq;
-  float half_angle = angle * 0.5;
-  float sin_half = sin(half_angle);
+  float half_angle = angle * 0.5f;
+  float sin_half = (float)sin(half_angle);
 
-  rq.w = cos(half_angle);
+  rq.w = (float)cos(half_angle);
   rq.x = x * sin_half;
   rq.y = y * sin_half;
   rq.z = z * sin_half;
